requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
fake-useragent==0.1.11
Flask==1.1.2
gevent==20.6.1
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
Pillow==7.1.2
PyMySQL==0.9.3
redis==3.5.3
requests==2.23.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
yarl==1.4.2
zope.event==4.4
zope.interface==5.1.0
